+++ src/app/nutrition/page.tsx
@@ -1,6 +1,5 @@
 import Image from 'next/image';
 import { TopGradient } from '../components/ui/TopGradient';
-import Header from '../components/ui/Header';
 
 const nutritionCardData: { header: string; desc: string; route: string }[] = [
   // {
@@ -27,9 +26,7 @@ const nutritionCardData: { header: string; desc: string; route: string }[] = [
 
 export default function Home() {
   return (
-    <main className="flex min-h-screen flex-col items-center justify-between p-24">
-      <Header h1="Nutrition" />
-
+    <>
       <TopGradient fromColor={'to-emerald-400'} toColor={'from-green-400'} />
 
       {/* 
@@ -55,6 +52,6 @@ export default function Home() {
           }
         )}
       </div> */}
-    </main>
+    </>
   );
 }
